+++ sdk/go/common/tokens/tokens_test.go
@@ -56,3 +56,53 @@ func TestTokens(t *testing.T) {
 	assert.Equal(t, p, modm.Module().Package().Name().String())
 	assert.Equal(t, p+TokenDelimiter+m+TokenDelimiter+mm, modm.String())
 }
+
+func TestTypeDisplayName(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		desc string
+		give Type
+		want string
+	}{
+		{
+			desc: "not enough parts",
+			give: "incomplete",
+			want: "incomplete",
+		},
+		{
+			desc: "no name",
+			give: "pkg:mod:",
+			want: "pkg:mod:",
+		},
+		{
+			desc: "no slash",
+			give: "pkg:mod:typ",
+			want: "pkg:mod:typ",
+		},
+		{
+			desc: "bad casing",
+			give: "pkg:Mod/foo:typ",
+			want: "pkg:Mod/foo:typ",
+		},
+		{
+			desc: "remove slash",
+			give: "pkg:mod/foo/bar:Bar",
+			want: "pkg:mod/foo:Bar",
+		},
+		{
+			desc: "remove up to last slash",
+			give: "pkg:mod/foo/bar/baz:Baz",
+			want: "pkg:mod/foo/bar:Baz",
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.desc, func(t *testing.T) {
+			t.Parallel()
+
+			assert.Equal(t, tt.want, tt.give.DisplayName())
+		})
+	}
+}
